Made reading of POSANG keyword from FITS file optional.
This is a temporary measure.

Datamodel changes, particularly to the pfsDesign and pfsConfig structures,
would mean that data used for test and development,
for example the drp_stella_data and drp_pfs_data repos would be unreadable
and would need to be regenerated.

To allow time and flexibility to coordinate such regeneration activities,
small changes such as the above are tempoarily set to be optional on read.
Once the test data are regenerated, such keywords are set to be
mandatory.